wr-servo: move generation of the state name earlier
Since commit f3d427ee, we lost the concept of "next_state" (to make code
simple), but this means that we should report to wr_mon the state string
we acted on, not the next. Otherwise, we see SYNC_PHASE all the time,
when we are actually waiting for SYNC_SEC or SYNC_NSEC to complete the
adjustment.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>