WR: set system time from WR time
The previous code to set system time from WR time is not really
working.

It is implemented in wrs_time_ops->set and it relies on ppsi being
non-wr slave before being wr slave, but this should not happen.  Such
code, with its "weird" message, remains in place to be used when we
run in non-wr mode (as slaves).

This new code is the right thing to do: when we are done syncing the
"big steps" in WR, and start dealing with sub-tick, we jump system
time.

There is no correction for TAI/UTC because the time operations for
arch-unux (set/get) already do the conversion.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>