wr-servo: run twice as fast
The old code used to discard every other tuple, because when "wait for
hardware was set", the code cleared the flag and did nothing. This was
exposed by one of the clean-up passes, but I chose not to change the
behaviour.

Now I do the normal work when wait_for_hw gets cleared.  However, as a
side effect, wr_mon was always saying "wait for hardware", because any
action would set the flag.

Thus, the solution is not setting the flag in TRACK_PHASE, but only
in the initial SYNC_SEC and SYNC_NSEC states. The code then checks
if the hardware is busy, irrespective of any flag.  The result in wr_mon
is matching reality, and the code works twice as fast as it used to.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>