general: cleanup prototypes
This adds -Wstrict-prototypes -Wmissing-prototypes and takes actions
accordingly.

Some of the solutions are ugly, especially in the bare architectures,
but I prefer to be warned if I take dangerous paths wrt prototypes while
writing further code.

I'm upset by those two not being included in -Wall, as they used to.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>