unix and wrs socket: open in non-blocking mode and report write errors
It may happen, though rarely, that frames are not sent (or not freed
at kernel level), and we build up an output queue, so at some point we
get -EAGAIN on sending.

With blocking writes, this situation was stalling the whole process.
This patch helped us diagnosing a problem in our own network driver,
which had a memory leakage, but using non-blocking sockets and
reporting erors is the right thing to do anyways.

This patch reports the error using pp_diag with a priority 0; thus
the message is always printed, but by virtue of pp_diag we have the
name of the port automatically added.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>