wrs-socket: use POLLERR for timestamps
While the manual pages say nothing, https://lwn.net/Articles/291793/
hints that we should use POLLERR instead of POLLIN.

I hoped to get rid of some of the poll messages, but nothing changes
in simple tests. I hope we get rid of the EAGAIN in the later recvmsg()
(likely with POLLIN we got the socket as readable, but nothing was
in the error queue -- we had a new incoming frame instead).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>