wrs-socket: decrease the timeout when waiting for a timestamp
Even under load, with queues full, the wrs delays frames less than one ms.

Thus, if we didn't get a timestamp in 1ms, we won't ever get it. Add
some software delay and clearly we don't need 100ms as in current
code.  Please note that all state machines are not running during this
wait, as we are single-process single-thread.

So now we wait for 6ms at most (3 times 2ms at most, but the first
time we may get an older timestamp, then discarded).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>