increase Announce rcv timeout to 20 - workaround for the WRS
It's done to prevent PPSi leaving Slave state on the WRS when high load of
background traffic is present. It's just a temporary workaround. In the future
the HDL of the switch has to be fixed so that PPSi frames have higher priority
than the normal traffic and are not dropped when WRS is overloaded.