wr-servo: TRACK_PHASE now monitors delays to zero it
During track-phase we used to find the setpoint and then track changes in
the master-to-slave delay, thus not zeroing any initial error that may
happen (e.g. the initial sample was more jittery than average).

Now we use the clock offset as correction source. Setpoint is
adjusted by 1/4th of the current offset, to smooth a little the jitter.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>