wrs: make tx timestamps work again with UDP
for tx stamps, when we added the check about the frame content
(to avoid problems with high data load, see 4179fcdc), it got broken
for UDP. That's because the cmesg we get back includes a raw frame
anyways, even if we transmitted UDP or other protocols.

So now check for the trailing part of the frame we get back, to
avoid any extra leading header whatever its lenght.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>