cmdline: remove -b option
-b didn't actually work. Even with a single interface and no config
file, the following error message was printed [arch-unix]:

Can not use -b option in multi-link conf

The bug was caused by a problem with the initialization value
of ppg->nlinks.

Removing this option looks wiser than trying to fix it, because
most users just work with a configuration file or -C on the command line,
so fixing -b would just increase the risk of breaking non-unix architectures
with no actual gain.

Signed-off-by: Davide Ciminaghi <ciminaghi@gnudd.com>