config: prepare for an e2e build with no p2p code inside
There is a size problem in wrpc-sw, when built with SNMP inside, so
some users, notably CERN, want to avoid the pdelay code (allegedly 7kB
in the binary).

Users who need pdelay don't actually run SNMP, so they don't have the
size problem.  This new setup allows a wrpc-sw without pdelay-related
code.  Wrpc with both mechanisms built in will be run-time configurable.

the wrs and unix configurations for peer delay are removed, because
for those architecture the choice is going to be performed at run time,
in the configuration file.

Please note that this commit is config-only, no code yet is there.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>