bugfix/workaround: unix-socket: disable settings for UDP/pdelay
This won't work, because packets must be sent to two different
IP multicast addresses. This is not supported.  The code being disabled
had the effect of setting the pdelay destination for all frames, even
for E2E runs.

This fixes E2E UDP operation, leaving P2P broken as it was.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>