arch-wrs: improve locking of shmem
Commit based on the commit from wrs-switch-sw described below.

userspace/libwr: improve locking of shmem

Keep explicitly lock bit in the sequence variable as a LSB bit.
This solution is backward compatible. The difference is only when shmem is
lock twice. Before it was treated as unlocked, now still as a lock.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>