trivial: remove unsed in/outbound_latency
While it may make sense to setup constant delay somewhere (for WR we
define them, for instance), the legacy implementation was unused and
wrong.

Unused because nobody sets it (even if we could do it in the config
file).  Wrong because inbound is never used, and both are global and
not per-port.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>