arch-wrs: in shmem, use pointers to structs intead of void pointers
Commit based on the commit from wrs-switch-sw described below.
Apply changes to the arch-wrs/wrs-startup.c to follow declaration change of
wrs_shm_alloc.

userspace/libwr: in shmem, use pointers to structs intead of void pointers

In shared memory functions take as parameters pinters to structures instead
of void pointers. By this compiler will warn when wrong type of pointer is
passed. For example when pointer to the data in shmem is passed instead of
pointer to the header.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>