remove arguments to net->send; some fixes as side effect
Most arguments of net->send are redundant, as the ppi includes all
info. In particular, chtype and "is_pdelay_addr" derive from the
message type.  And the timestamp is always internal.

I chose not to remove message, len and msgtype because the message
being sent includes the header (ppi->tx_offset), so there's some
calculation to extract them.  This may happen in a later commit, if I
find that is a size improvement.

Bugs fixed as a side effect:
   - the "if" for "is_pelay" missed PDELAY_RESP_F_UP (now it's in msgtype.c)
   - wrpc pdelay was always using the e2e mac address

This commit decreases the size of all archs by 60 bytes, but increases
wrpc by 20 bytes because of the fix above

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>