bugfix: wrs was crazy when slave of non-wr masters
The main problem was introduced by:

   d2a57023 std-servo: fix initial jump for p2p slaves

There, we falled back on unix->get_time to make adjustments if
hal->get_time failed.  Only happens as non-wr slave.

Actually, the get_time call has never been implemented in the hal, and
we always falled on unix time. But if frames are timestamped with
WR time, we can adjust based on Unix time.

We now read WR time internally, waiting for something better (we have
constant addresses in here).

Note: we are still unable to steer frequency, despite the efforts
(and it used to work, ages ago).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>