proto-ext-whiterabbit: change the order of fields in struct wr_servo_state
In wrpc's SNMP we store an offset between beginning of a structure and
particullar field. To save memory we store this offset as a uint8_t, which
limits the offset of a filed to 255 bytes. It was necessary to move following
fields below 255 offset:
--n_err_state
--n_err_offset
--n_err_delta_rtt

Compiling wrpc-sw with an offset greater than 255, a compiler prints warnings
like:
lib/snmp.c:405:2: warning: large integer implicitly truncated to unsigned type

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>