general fix: implement SYNCHRONIZATION_FAULT
If we stopped sending to the master or the peer (for traffic or
whatever; in my case with "fault drop"), we wouldn't notice the
problem.

This looks like SYNCHRONIZATION_FAULT (9.2.6.12), so this reuses the
almost-unused TO_FAULTY, renaming it to a more generic TO_FAULT.

9.2.6.12 says we should reach uncalibrated, but since uncalibrated doesn't
exits (it is never entered, it's dead and untested code at this point),
I handle the problem just like the timeout receiving announce messages.

For wr, I reset the servo, so the problem can be seen.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>