wr/wrpc: implement absolute calibration support
This is a port of previous work by Peter Jansweijer from nikhef.

To perform absolute calibration, we need a grand-master look-alike
mode that sends sync once a second (and hopefully slightly after the
pps signal).

Using a special gateware that sends a pulse whenever a frame is
transmitted and received, users can correlate collected timestamps
(T1 and T4), this special pulse and the pps pulse of the node.

The procedure for absolute calibration is described in

http://www.ohwr.org/attachments/4542/WhiteRabbitAbsoluteCalibrationProcedure.pdf

Another commit, in wrpc-sw, adds "mode abscal" for this feature to be used.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>