Soft PLL returns unlock state sometime
This was due to the introduction of system timers in the main loop of
the switch. In minirpc the pool() system call exited sometime when the
timer fired. Some other problems have been found during the debugging
process. Some of them have been already adressed :
- In minirpc, loop on pool() system call if the call is interrupted.
- Wrong check in function wrs_locking_poll(). Was working by chance.
- L12_state_machine() was calling to many often the state handle. It was
increasing the number of hardware calls for nothing.

TODO: The returned values for the function wrs_locking_poll(), halexp_lock_cmd(), and hal_port_check_lock() should be reviewed to be more consistent.