Improvee detection of multiple PPSi running instances
When /etc/init.d/ppsi.sh is called with the parameter 'restart', it
happens sometimes that the restarted PPSI process detects that another
one is running but it is not really the case. To  sove this problem, the
PPSI process is now  checking  few times with a delay in between before
to decide to exit.