New field activePeer in pp_instance structure
This field is used to store the active peer connected to an instance. In
wr_mon tool, it replaces the field peer on the display which was not
very explicit when many peers are connected on the same port. The field
activePeer takes the value of peer field in the following cases :

           MASTER          UNCALIBRATED/SLAVE
P2P     Rx PDELAY_RESP    Rx qualified ANNOUNCE
          (Filtered)
E2E       DELAY_RESP      Rx qualified ANNOUNCE

In other cases, the activePeer field is set to 0