general: include <limits.h>
This version of ppsi, currently submodule of proposed_master of
wr-switch-sw, is not building because of INT_MAX, UINT_MAX and the like.

This includes <limits.h> in all files where this is needed.

Obviously, <limits.h> will be missing in wrpc-sw and bare builds, but
this version does not build for most architectures anyways (including
arch-unix).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>