WR extension: Change how a new parent is detected
The detection of a new parent has been moved in a call back of the S1
treatment. The previous implementation had some side effect due to the
fact that the detection was coming too late in some cases (pb with
erbest calculation)

The doRestart called by the servo is now extension specific. It help to
make a more cleaner restart.