tools/dump_mem_ppsi_wrpc: fix printout of pp_frgn_master
Fix the printout of pp_frgn_master if more than 1 foreign master is registered.
Before the size of struct pp_frgn_master was calculated based on the size on
a host system, not on LLM32. In practice the size was different by 4 bytes.

This commit requires the following commit from WRPC repo:
wrpc-dump: add dump of task list

Signed-off-by: Adam Wujek <adam.wujek@creotech.pl>