[BUG: #29] copy activePeer when in slave mode
Before this commit activePeer was filled when the siwtch was in master
mode or when the external port configuration was active.
This commit adds storing the MAC of each foreign master, the MAC of
the chosen foreign master is copied to activePeer.

IMHO: activePeer (or peer in general) should not be the MAC but
the portIdentity of a port on the master.

Signed-off-by: Adam Wujek <adam.wujek@creotech.pl>