arch-wrs: relicensing mini-rpc to LGPL to fix ppsi compatibility
After a discussion with CERN, it has been agreed to allow a relicensing
of mini-rpc source code to LGPL v2.1 in order to fix LGPL license for
ppsi when compiled for arch-wrs (the GPL of mini-rpc was breaking the
intention to realase PPSi as LGPL for arch-wrs).

The source code of mini-rpc is fully under CERN copyright with some
fixes from Seven Solutions, and all parties agree for this relicensing.